refactor: move pollingTask
