+++ packages/maskbook/src/plugins/Wallet/apis/coingecko.ts
@@ -1,4 +1,4 @@
-import { pollingTask } from '../../../utils'
+import { pollingTask } from '@masknet/shared'
 import { currentEtherPriceSettings } from '../settings'
 
 const ETH_PRICE_POLLING_DELAY = 30 /* seconds */ * 1000 /* milliseconds */
@@ -11,7 +11,7 @@ interface TokenRecord {
     [token: string]: PriceRecord
 }
 
-async function fetchTokenPrice(token = 'ethereum', currency = 'USD') {
+async function fetchTokenPrice(token = 'ethereum', currency = 'usd') {
     const response = await fetch(`https://api.coingecko.com/api/v3/simple/price?ids=${token}&vs_currencies=${currency}`)
     const data = (await response.json()) as TokenRecord | null
     if (!data) return 0
@@ -21,7 +21,7 @@ async function fetchTokenPrice(token = 'ethereum', currency = 'USD') {
 export function trackEtherPrice() {
     const { reset } = pollingTask(
         async () => {
-            currentEtherPriceSettings.value = await fetchTokenPrice('ethereum', 'USD')
+            currentEtherPriceSettings.value = await fetchTokenPrice('ethereum', 'usd')
             return false
         },
         {
