+++ packages/maskbook/src/utils/utils.ts
@@ -128,44 +128,6 @@ export function batchReplace(source: string, group: Array<[string | RegExp, stri
     return storage
 }
 
-export function pollingTask(
-    task: () => Promise<boolean>,
-    {
-        autoStart = true,
-        delay = 30 * 1000,
-    }: {
-        autoStart?: boolean
-        delay?: number
-    } = {},
-) {
-    let canceled = !autoStart
-    let timer: NodeJS.Timeout
-
-    const runTask = async () => {
-        if (canceled) return
-        let stop = false
-        try {
-            stop = await task()
-        } catch (e) {
-            console.error(e)
-        }
-        if (!stop) resetTask()
-    }
-    const resetTask = () => {
-        canceled = false
-        clearTimeout(timer)
-        timer = setTimeout(runTask, delay)
-    }
-    const cancelTask = () => {
-        canceled = true
-    }
-
-    if (!canceled) runTask()
-    return {
-        reset: resetTask,
-        cancel: cancelTask,
-    }
-}
 export function addUint8Array(a: ArrayBuffer, b: ArrayBuffer) {
     const x = new Uint8Array(a)
     const y = new Uint8Array(b)
