+++ packages/shared-base/src/utils/index.ts
@@ -2,4 +2,5 @@ export * from './parseURL'
 export * from './delay'
 export * from './timeout'
 export * from './enum'
+export * from './pollingTask'
 export * from './sessionStorageCache'
